Improve shader sending method to GAL, use a memory interface instead of reading a fixed array size and sending every time

+++ Ryujinx.Graphics/Gal/Shader/ShaderDecodeFlow.cs
@@ -15,7 +15,7 @@ namespace Ryujinx.Graphics.Gal.Shader
                 throw new NotImplementedException();
             }
 
-            int Target = ((int)(OpCode >> 20) << 8) >> 8;
+            long Target = ((int)(OpCode >> 20) << 8) >> 8;
 
             Target += Block.Position + 8;
